UzmanWork/logistics-app/pyproject.toml
[tool.poetry]
authors = ["Nedy88 <nedy88@gmail.com>"]
description = ""
name = "logistics-app"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
PyJWT = {extras = ["crypto"], version = "^2.6.0"}
SQLAlchemy = {extras = ["mypy"], version = "^1.4.42"}
aio-pika = "^8.2.5"
aiodns = "^3.0.0"
aiofiles = "^23.1.0"
aiohttp = "^3.9.4"
alembic = "^1.9.3"
asyncpg = "^0.26.0"
beautifulsoup4 = "^4.11.1"
black = "24.3.0"
boto3 = "^1.25.5"
celery = {extras = ["gevent"], version = "^5.2.7"}
celery-types = "^0.15.0"
email-validator = "^1.3.0"
fastapi = "0.109.2"
ffmpeg-python = "^0.2.0"
flower = "^1.2.0"
greenlet = "^2.0.0"
gunicorn = "^22.0.0"
isort = "^5.10.1"
numpy = "^1.23.5"
openai = "^0.27.6"
phonenumbers = "^8.13.3"
progiter = "^1.0.1"
pytest-alembic = "^0.10.0"
pytest-asyncio = "^0.20.2"
pytest-postgresql = "^4.1.1"
python = "^3.10"
python-dotenv = "^0.21.0"
requests = "^2.28.1"
sendgrid = "^6.9.7"
types-requests = "^2.28.11.2"
uvicorn = "^0.18.3"
websockets = "^10.4"
# TODO (oliverscheel): this breaks on MacOS, so we skip it for now
# -> you can't run DB unit tests on it.
google-cloud-aiplatform = "^1.26.0"
orjson = "^3.9.15"
psycopg = {extras = ["c"], version = "^3.0.15", markers = "sys_platform == 'linux'"}
pyamqp = "^0.1.0.7"
python-dateutil = "^2.8.2"
pytz = "^2022.7.1"
redis = {extras = ["hiredis"], version = "^4.5.5"}
sentry-sdk = "^1.40.4"
slack-sdk = "^3.19.5"
twilio = "^7.16.0"
types-aiofiles = "^22.1.0.7"
types-python-dateutil = "^2.8.19.8"
types-pytz = "^2022.7.1.0"
stamina = "^23.1.0"
cryptography = "^42.0.4"
tornado = "^6.3.3"
certifi = "^2023.7.22"
pydantic = "~1.10.13"
pytest-xdist = "^3.3.1"
py-spy = "^0.3.14"
aiobotocore = "^2.12.2"
python-multipart = "^0.0.9"
filetype = "^1.2.0"

[tool.poetry.group.dev.dependencies]
faker = "^15.3.4"
flake8 = "^5.0.4"
httpx = "^0.23.3"
mypy = "^1.6.0"
openapi-python-client = "^0.14.1"
pytest = "^7.2.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.10.0"
types-redis = "^4.5.5.2"

[tool.isort]
profile = "black"
skip = ["api_client", "frontend"]

[tool.black]
skip_magic_trailing_comma = true
preview = true

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.mypy]
explicit_package_bases = true
namespace_packages = true
plugins = ["sqlalchemy.ext.mypy.plugin", "pydantic.mypy"]
python_version = "3.10"
show_error_codes = true
strict = true
warn_return_any = true
warn_unused_configs = true
exclude = ["frontend"]

[[tool.mypy.overrides]]
ignore_missing_imports = true
module = [
  "pytest",
  "bs4",
  "boto3",
  "botocore",
  "botocore.*",
  "aiobotocore.*",
  "progiter",
  "sendgrid.*",
  "twilio.*",
  "python_http_client.*",
  "ffmpeg.*",
  "shapely.*",
  "vertexai.*",
  "filetype.*",
]

[[tool.mypy.overrides]]
ignore_missing_imports = true
implicit_reexport = true
module = [
  "sqlalchemy",
  "sqlalchemy.*",
]

[tool.pytest.ini_options]
addopts = "--tb=short"
asyncio_mode = "auto"
filterwarnings = ["ignore::pytest.PytestCacheWarning", "ignore::DeprecationWarning"]
